requirements.txt
alabaster==0.7.12
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
beautifulsoup4==4.7.1
setuptools==58.0.4
certifi==2019.3.9
chardet==3.0.4
docutils==0.14
idna==2.8
imagesize==1.1.0
Jinja2==2.10.1
lxml==4.3.0
MarkupSafe==1.1.1
more-itertools==7.0.0
nyawc==1.8.2
packaging==19.0
pdfminer3k==1.3.4
Pillow==9.5.0
pluggy==0.9.0
ply==3.11
pockets==0.7.2
py==1.8.0
Pygments==2.3.1
pyparsing==2.4.0
pytest==4.4.0
python-docx==0.8.10
pytz==2019.1
requests==2.21.0
requests-file==1.4.3
requests-toolbelt==0.9.1
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.9.1
Sphinx==1.8.3
sphinx-better-theme==0.1.5
sphinxcontrib-napoleon==0.7
sphinxcontrib-websupport==1.1.0
tldextract==2.2.1
urllib3==1.24.1
xlrd==1.2.0
